test: restore manual open urls file
